Attempt to remove line wrap from new forum topics posted via email. Adds new function - bp_rbe_remove_line_wrap_from_plaintext().
This is hooked to the 'bp_rbe_parse_email_body_new' filter.  It's a little hacky.  Will leave it in for now until people complain :)

Also change the priority of the 'imap_delete' function on the 'bp_rbe_imap_loop' hook to 99 so this is done later in the firing process.