Add failure message to sender if user isn't part of group. See #12.
If a user decides to manually leave a group, but decided to reply to an older RBE group forum item, this forum reply item will get blocked from posting, but a failure message isn't sent back to the sender to inform them of what happened.

This commit sends back a failure message to the sender to inform him/her that their message was not posted because they are no longer a member of the group.

In bp-rbe-classes.php:
* Separate group member check and group member banned check for failure message purposes.
* Drop support for super admins who aren't a member of a group to reply to group forum posts by email.  This didn't make sense to begin with.

In bp-rbe-functions.php:
* Introduce bp_rbe_get_group_member_info() to get all info about a group member in one DB query.
* In bp_rbe_imap_log_no_matches(), add internal log and failure messages for users who are not a part of a group anymore.