Do not show the 'Post New Topics via Email' block if not a member.
Previously, if a user was logged in and was viewing a group they were not a member of, that user could still view the 'Post New Topics via Email' block.  This has been corrected in this commit.

Also, this blocks super admins who are not a member of the group from viewing this block as well as it doesn't make much sense to me.

Fixes #17.