Introduce locking mechanism when connecting to the IMAP server.
On occasion, RBE can try to connect to the inbox twice within the same
minute. This can create situations where double-posting inadvertenly
occurs.

To address this, in this commit, we introduce a transient - 'bp_rbe_lock'
 - and check it before we try to connect to the inbox.

If we're already connected and if we're attempting to connect for a second
time, RBE will stop the second connection attempt.

The transient expires after 60 seconds, which should be adequate time to
check if imap_open() has completed its attempt.