For non-RBE emails, add a line above each email message stating that you should not reply to this email.
This is to prevent the recipient from replying to non-RBE emails, which
will not get parsed by RBE.

Note: This line will not show up on activation emails and when an email is
sent from the WP admin dashboard.

All other emails will add the line.

eg. friendship request emails, group member promotion emails, and any other
email that is sent from the frontend.

This can be disabled with the 'bp_rbe_show_non_rbe_notice' filter if needed.