In the inbox loop, before getting the userdata, flush the object cache if the default object cache is in use.
This is done because the default object cache is only meant to be used on a
single page load.  Since RBE runs in the background, we need to flush the
object cache so WP will do a direct DB query the next time it fetches any
data.

To check if the default object cache is in use, this commit introduces the
bp_rbe_is_external_object_cache_used() function.

See #28.