Legacy forums: Escape forum text when checking for duplicate replies.
bp_forums_reply_exists() does not escape the text before the post text
in the database.  This can lead to database errors when using an
apostrophe in the body of the forum reply.

This commit escapes the text using WP's esc_sql() before using
the bp_forums_reply_exists() function to address this issue.