Replace WP cron with custom version.
In this commit, we're removing integration with WordPress cron due to edge-
case scheduling issues and creating our own version of cron using
transients.

The side-effect of this is we are also able to automatically reconnect to
the inbox after an inbox check with the keep-alive value expires.  The next
commit will add an admin option to disable auto-reconnection.

See #16.