Add autoconnect option.
In commit 006f3b2, we overhauled cron for our custom version.  This led to
the ability to automatically reconnect to an inbox after the keep-alive
value expires.

However, if we wanted to disable this feature, there was no admin option to
do so.

This commit adds an admin option to disable the autoconnect feature.

See #16.