Replace transients with options saved as timestamps.
In commit 006f3b224b, transients replaced the older method of using WP-cron.
However, I was using transients more as an expiration date.

This is the incorrect way to use transients as transients uses the object
cache when a persistent object cache is in use.

A persistent object cache is a cache and not a data store.  Meaning that the
transient date guarantees the data will not exist past the expiration time.
However, the transient *can* expire earlier than this date, which could lead
to inaccuracies and potential double-connections.

Switching to options resolves this issue.

See #16.