Set mode to 'imap' if it was previously being used prior to dual mode.
RBE now supports two modes, however for those that were using RBE prior to
this, we need to set the mode to 'imap' to ensure the admin doesn't have
to visit the RBE settings page and resave the settings.

This occurs during plugin activation.  So RBE needs to be deactivated and
reactivated for this to take effect.