Fix fatal error in BP_Reply_By_Email_Parser::get_header() when using Mandrill and sending a failure feedback message to the sender.
Mandrill uses an object for its email headers, while RBE's internal IMAP
class uses an array for its headers.  Since the get_header() method only
looked for an array, this caused a "Cannot use object of type stdClass as
array" fatal error.

This commit type casts the $headers parameter so it is always an array,
which fixes this problem.