bbPress: Ensure RBE author stays subscribed to topic.
When a user replies to a non-BuddyPress topic, we need to explicitly
tell bbPress to keep the topic subscription for the user.  Otherwise, the
user will be unsubscribed after posting by email.

This commit sets the $_POST['bbp_topic_subscription'] flag so bbPress
see this.  This is due to how bbp_update_reply() works.

See #24.