bbPress: Automatically provision meta caps when replying via RBE.
When viewing the frontend, bbPress dynamically provisions caps based on the
logged-in user, however when a user is replying via RBE, that user is not
logged in.

To abide by bbPress' meta caps filter - 'bbp_map_meta_caps' - we now ensure
that we are using this filter as well.

See #24.