bbPress: Change failure message to sender to add the reply at the bottom of the email.
Previously, if a user posts a long new topic or reply that fails, that user
would not be able to see the type of RBE error until the very end of the
email.

Now, the type of error message is placed at the beginning of the email and
a copy of the attempted reply is added at the end.