bbPress: Remove flood check.
For the flood check, bbPress takes the user's IP and uses the last posted
time from that IP address.  The problem with this is RBE will always use
the default IP address assigned by bbPress (127.0.0.1), so it's quite
possible that conflicts will occur for users attempting to post via RBE.

This commit removes this bbPress feature.

Fixes #66.