IMAP: Do not load inbox spawner if auto-connect is on.
If auto-connect is off, we'll rely on the old "page load to spawn inbox
check" method.

If auto-connect is on, we're going to be using the auto-connect admin option
to manually ping the IMAP inbox from the admin settings page.  This will
prevent duplicate IMAP connections from occurring and will also workaround
using wp_remote_post() to spawn the inbox check.  This logic will be
added in subsequent commits.

See #40.