Admin: Stop IMAP AJAX response from hanging by clearing the current session.
This only occurs if a 3rd-party plugin is opening a PHP session with
session_start() without closing the session properly.  What happens if the
session isn't closed is all subsequent AJAX responses and browser requests
will hang since our IMAP inbox check is still running in the background.

Thanks to this Stack Overflow article for helping me out:
http://stackoverflow.com/a/9906646

If you wrote a plugin that does this, you are forever on my naughty list.
Thanks for wasting my precious time!

See #40.