IMAP: If auto-connect is enabled, send an email to the site admin if connection is down.
When auto-connect is enabled, there are occasions where the IMAP connection
can be disconnected such as if GMail is acting up or if the server that
WordPress is hosted on is down.  When this happens, the IMAP connection
needs to be reinitiated from the RBE admin settings page since IMAP
connections can only be initiated from the admin area when the auto-connect
option is on.

To counteract this, this commit adds an hourly check to see if the IMAP
connection is active.  If not connected, an email is sent to the site admin
by default to prompt the person to re-enable the IMAP connection on the RBE
admin settings page.  A filter exists to change the email recipients.

See #40.