Fix inaccurate PHPMailer 'Date' timestamp when auto-connect is on and when using IMAP mode.
BuddyPress uses a static version of PHPMailer:
https://buddypress.trac.wordpress.org/browser/tags/2.8.2/src/bp-core/classes/class-bp-phpmailer.php?marks=28,39,47#L27

When RBE is using IMAP mode with auto-connect turned on, BP's PHPMailer
would never change the 'Date' header from when it was first initialized.

This meant that subsequent emails would all use the older timestamp, which
is wrong!

This commit remedies this by always setting the 'Date' field when IMAP mode
and auto-connect is on.

See #23.