IMAP: Ensure our cron event is scheduled when auto-connect is enabled.
Previously, our cron event was only scheduled when connecting to the IMAP
inbox on the initial connection.  However, if our IMAP connection ever
dies, our cron event is never rescheduled, so admins would never know if
the IMAP connection needs to be restarted or not.

This commit always checks to register our cron event to ensure admins can
restart the IMAP connection when needed.

See #40.