bbPress: Parse IMAP email attachments and save temporarily.
What we're doing here is:
- Checking to see if the GD bbPress Attachments plugin is active;
- Checking if the current reply-by-email item is a reply to a bbPress post;
- Using the logic from GD bbPress Attachments to check if the bbPress forum allows attachments; and
- Using the logic from GD bbPress Attachments to check if the attachment is too large

At this juncture, RBE hasn't parsed the email yet, so what we do here if
we've passed all the checks above is save the attachments to the server's
temporary directory and pass along the the attachment's metadata (filepath
to attachment saved in temporary directory and the attachment's filename)
to the RBE parser.

We can only insert our attachments after RBE has parsed the email and
created the new bbPress post.  We will do this in the next commit.

See #104.