bbPress: Add attachments parsed from email message into the bbPress reply.
In commit ebd511b, we saved any attachments that were added to a bbPress
reply email to a temporary directory, but had to wait until the actual
bbPress post was created to insert the attachment.

This commit does the actual attachment insertion and utilizes WP's
media_handle_sideload() function to do add our attachments to the bbPress
post ID.

See #104.