Better method in removing trailing breaklines in BuddyPress HTML template.
When RBE adds its marker to the BuddyPress HTML template, it adds the
marker plus two '\n\n' calls.  In BuddyPress 3.0, the marker is wrapped
around a '<p>' tag, which needs to be removed.  In older versions of
BuddyPress, the '\n\n' call gets converted to two '<br />'.

Previously, commit 8a5922b was too aggressive in stripping this markup.

This commit does the following:

- For BuddyPress 3.0, only remove the marker and the paragraph tag.
  Also, strips a trailing <br> for aesthetic reasons (see BP ticket 7989).
- For older versions of BuddyPress, only remove the two trailing '<br />'
  calls.

Fixes #117.