BP Groupblog: Support groupblog's comment moderation setting.
If the groupblog has comment moderation turned on (the "Before a comment
appears, comment must be manually approved" option is checked in the WP
admin dashboard's "Settings > Discussion" page) and a group member replies
to a groupblog post via email, the email reply is now placed in the
groupblog's comment moderation queue.  Previously, the comment was
automatically published, which isn't desired behavior.

Fixes #121.