Refactor how the 'Reply-To' email header is set with PHPMailer, again.
Previously, commits f297579 and f7831e8 added a workaround for a
PHPMailer bug where if the Reply-To email header is too long, it would
be silently discarded. The workaround uses the custom headers method
to add the 'Reply-To' header.

In WordPress 5.5, PHPMailer was updated to v6.1.6 from v5.2.27.
PHPMailer's custom headers method now encodes all custom header
values, which can corrupt the 'Reply-To' header if added this way.

To address this for PHPMailer v6, we will now go back to using
PHPMailer's internal addReplyTo() method, but we set the email address
validator to use the more lenient 'html5' method instead of 'php' or
'pcre' before we add the 'Reply-To' header.

Fixes #123.